Reorganized structure and simplified goap implementation to be more straight forward with godot.
