+++ .idea/.idea.Godot-GOAP-AI/.idea/libraries/GdSdk.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="GdSdk" type="GdScript">
+    <CLASSES />
+    <JAVADOC />
+    <SOURCES>
+      <root url="file://$APPLICATION_HOME_DIR$/plugins/rider-gdscript/sdk/extracted/4.4.1" />
+    </SOURCES>
+  </library>
+</component>
