--- .idea/.idea.Godot-GOAP-AI/.idea/libraries/GdSdk_Master.xml
@@ -1,10 +1,0 @@
-<component name="libraryTable">
-  <library name="GdSdk Master" type="GdScript">
-    <properties path="C:\Users\alexs\AppData\Local\JetBrains\Rider2024.3\projects\.idea.godot-goap-ai.793818bd\sdk\GdSdk Master" version="Master" date="2024-06-01T15:14:16.000+02:00" />
-    <CLASSES />
-    <JAVADOC />
-    <SOURCES>
-      <root url="file://$USER_HOME$/AppData/Local/JetBrains/Rider2024.3/Projects/.idea.godot-goap-ai.793818bd/sdk/GdSdk Master" />
-    </SOURCES>
-  </library>
-</component>
