--- Source/GOAP/Godot/Actions/PickUpItem/GoapActionPickUpItem.cs
@@ -1,6 +1,0 @@
-namespace GodotGOAPAI.Source.GOAP.Godot.Actions.PickUpItem;
-
-public class GoapActionPickUpItem
-{
-    
-}
