--- Source/GOAP/Godot/CustomResource/GodotGoapResourceType.cs.uid
@@ -1,1 +1,0 @@
-uid://bwcool4t1d8ms
