--- Source/GOAP/WorldStateGenerator/IGoapWorldStateGenerator.cs
@@ -1,8 +1,0 @@
-﻿using GodotGOAPAI.Source.GOAP.Models;
-
-namespace GodotGOAPAI.Source.GOAP.WorldStateGenerator;
-
-public interface IGoapWorldStateGenerator<T, TReturn>
-{
-    GoapWorldStateModel<TReturn> GenerateWorldStateModel(T worldCollectionsRootNode);
-}
