+++ Source/Goap/Actions/Abstraction/GoapActionBase.cs
@@ -1,6 +1,6 @@
 ﻿using Godot;
 
-namespace GodotGOAPAI.Source.GOAP.Godot.Actions.Abstraction;
+namespace GodotGOAPAI.Source.Goap.Actions.Abstraction;
 
 public abstract class GoapActionBase : IGoapAction
 {
