+++ Source/Goap/Actions/Pick Up Item/GoapActionPickUpItem.cs
@@ -1,0 +1,6 @@
+namespace GodotGOAPAI.Source.Goap.Actions.Pick_Up_Item;
+
+public class GoapActionPickUpItem
+{
+    
+}
