+++ Source/Goap/Custom Resource/GoapResource.cs
@@ -1,10 +1,10 @@
 ﻿using Godot;
 using GodotGOAPAI.Source.GOAP.Models;
 
-namespace GodotGOAPAI.Source.GOAP.Godot.CustomResource;
+namespace GodotGOAPAI.Source.Goap.Custom_Resource;
 
 [GlobalClass]
-public partial class GodotGoapResource : Resource
+public partial class GoapResource : Resource
 {
     [Export(PropertyHint.Enum)] 
     public GoapResourceType GoapResourceType { get; private set; }
