+++ Source/Goap/Custom Resource/GoapResource.cs.uid
