+++ Source/Goap/Models/GoapResourceType.cs.uid
