+++ Source/Goap/Models/GoapWorldStateModel.cs
