+++ Source/Goap/Models/IGoapItem.cs
